# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TAG - Tactical Adaptation Group API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface RegisterBody {
  username: string;
  email: string;
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface AuthUser {
  id: number;
  username: string;
  email: string;
  role: string;
  status: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  discordTag: string | null;
  createdAt: string;
}

export interface UserProfile {
  id: number;
  username: string;
  email: string;
  role: string;
  status: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  discordTag: string | null;
  createdAt: string;
}

export interface UpdateRoleBody {
  role: string;
}

export interface UpdateStatusBody {
  status: string;
}

export interface MessageItem {
  id: number;
  senderId: number;
  recipientId: number;
  senderUsername: string;
  recipientUsername: string;
  subject: string;
  body: string;
  isRead: boolean;
  createdAt: string;
}

export interface SendMessageBody {
  recipientId: number;
  subject: string;
  body: string;
}

export interface StaffApplicationBody {
  gamertag: string;
  games: string[];
  experience: string;
  motivation: string;
}

export interface StaffApplicationItem {
  id: number;
  userId: number;
  username: string;
  gamertag: string;
  games: string[];
  experience: string;
  motivation: string;
  status: string;
  /** @nullable */
  reviewNote: string | null;
  /** @nullable */
  reviewedBy: number | null;
  createdAt: string;
}

export interface ReviewApplicationBody {
  status: string;
  reviewNote?: string;
}

export type ListUsersParams = {
  role?: string;
  status?: string;
};