# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TAG - Tactical Adaptation Group API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication
  - name: users
    description: User management
  - name: messages
    description: Inbox messaging
  - name: staff-applications
    description: Staff applications
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterBody"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email or username already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Log out
      responses:
        "204":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (moderator/admin only)
      parameters:
        - in: query
          name: role
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserProfile"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{id}/role:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /users/{id}/status:
    patch:
      operationId: updateUserStatus
      tags: [users]
      summary: Update user status (moderator/admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStatusBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /messages/inbox:
    get:
      operationId: getInbox
      tags: [messages]
      summary: Get inbox messages
      responses:
        "200":
          description: Inbox
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /messages/sent:
    get:
      operationId: getSent
      tags: [messages]
      summary: Get sent messages
      responses:
        "200":
          description: Sent messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MessageItem"

  /messages:
    post:
      operationId: sendMessage
      tags: [messages]
      summary: Send a message
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageBody"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageItem"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /messages/{id}/read:
    patch:
      operationId: markMessageRead
      tags: [messages]
      summary: Mark a message as read
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageItem"

  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /staff-applications:
    post:
      operationId: submitStaffApplication
      tags: [staff-applications]
      summary: Submit a staff application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StaffApplicationBody"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffApplicationItem"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: listStaffApplications
      tags: [staff-applications]
      summary: List staff applications (moderator/admin)
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffApplicationItem"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /staff-applications/{id}:
    patch:
      operationId: reviewStaffApplication
      tags: [staff-applications]
      summary: Approve or reject a staff application (moderator/admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewApplicationBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffApplicationItem"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error

    RegisterBody:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
      required:
        - username
        - email
        - password

    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password

    AuthUser:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        bio:
          type: ["string", "null"]
        discordTag:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - role
        - status
        - bio
        - discordTag
        - createdAt

    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
        status:
          type: string
        bio:
          type: ["string", "null"]
        discordTag:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - username
        - email
        - role
        - status
        - bio
        - discordTag
        - createdAt

    UpdateRoleBody:
      type: object
      properties:
        role:
          type: string
      required:
        - role

    UpdateStatusBody:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    MessageItem:
      type: object
      properties:
        id:
          type: integer
        senderId:
          type: integer
        recipientId:
          type: integer
        senderUsername:
          type: string
        recipientUsername:
          type: string
        subject:
          type: string
        body:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - senderId
        - recipientId
        - senderUsername
        - recipientUsername
        - subject
        - body
        - isRead
        - createdAt

    SendMessageBody:
      type: object
      properties:
        recipientId:
          type: integer
        subject:
          type: string
        body:
          type: string
      required:
        - recipientId
        - subject
        - body

    StaffApplicationBody:
      type: object
      properties:
        gamertag:
          type: string
        games:
          type: array
          items:
            type: string
        experience:
          type: string
        motivation:
          type: string
      required:
        - gamertag
        - games
        - experience
        - motivation

    StaffApplicationItem:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        username:
          type: string
        gamertag:
          type: string
        games:
          type: array
          items:
            type: string
        experience:
          type: string
        motivation:
          type: string
        status:
          type: string
        reviewNote:
          type: ["string", "null"]
        reviewedBy:
          type: ["integer", "null"]
        createdAt:
          type: string
      required:
        - id
        - userId
        - username
        - gamertag
        - games
        - experience
        - motivation
        - status
        - reviewNote
        - reviewedBy
        - createdAt

    ReviewApplicationBody:
      type: object
      properties:
        status:
          type: string
        reviewNote:
          type: string
      required:
        - status