# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TAG - Tactical Adaptation Group API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Register a new member
 */
export const RegisterBody = zod.object({
  username: zod.string(),
  email: zod.string(),
  password: zod.string(),
});

/**
 * @summary Log in
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  email: zod.string(),
  role: zod.string(),
  status: zod.string(),
  bio: zod.string().nullable(),
  discordTag: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary Get current user
 */
export const GetMeResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  email: zod.string(),
  role: zod.string(),
  status: zod.string(),
  bio: zod.string().nullable(),
  discordTag: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary List all users (moderator/admin only)
 */
export const ListUsersQueryParams = zod.object({
  role: zod.coerce.string().optional(),
  status: zod.coerce.string().optional(),
});

export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  email: zod.string(),
  role: zod.string(),
  status: zod.string(),
  bio: zod.string().nullable(),
  discordTag: zod.string().nullable(),
  createdAt: zod.string(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Get a user by ID
 */
export const GetUserParams = zod.object({
  id: zod.coerce.number(),
});

export const GetUserResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  email: zod.string(),
  role: zod.string(),
  status: zod.string(),
  bio: zod.string().nullable(),
  discordTag: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary Update user role (admin only)
 */
export const UpdateUserRoleParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateUserRoleBody = zod.object({
  role: zod.string(),
});

export const UpdateUserRoleResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  email: zod.string(),
  role: zod.string(),
  status: zod.string(),
  bio: zod.string().nullable(),
  discordTag: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary Update user status (moderator/admin)
 */
export const UpdateUserStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateUserStatusBody = zod.object({
  status: zod.string(),
});

export const UpdateUserStatusResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  email: zod.string(),
  role: zod.string(),
  status: zod.string(),
  bio: zod.string().nullable(),
  discordTag: zod.string().nullable(),
  createdAt: zod.string(),
});

/**
 * @summary Get inbox messages
 */
export const GetInboxResponseItem = zod.object({
  id: zod.number(),
  senderId: zod.number(),
  recipientId: zod.number(),
  senderUsername: zod.string(),
  recipientUsername: zod.string(),
  subject: zod.string(),
  body: zod.string(),
  isRead: zod.boolean(),
  createdAt: zod.string(),
});
export const GetInboxResponse = zod.array(GetInboxResponseItem);

/**
 * @summary Get sent messages
 */
export const GetSentResponseItem = zod.object({
  id: zod.number(),
  senderId: zod.number(),
  recipientId: zod.number(),
  senderUsername: zod.string(),
  recipientUsername: zod.string(),
  subject: zod.string(),
  body: zod.string(),
  isRead: zod.boolean(),
  createdAt: zod.string(),
});
export const GetSentResponse = zod.array(GetSentResponseItem);

/**
 * @summary Send a message
 */
export const SendMessageBody = zod.object({
  recipientId: zod.number(),
  subject: zod.string(),
  body: zod.string(),
});

/**
 * @summary Mark a message as read
 */
export const MarkMessageReadParams = zod.object({
  id: zod.coerce.number(),
});

export const MarkMessageReadResponse = zod.object({
  id: zod.number(),
  senderId: zod.number(),
  recipientId: zod.number(),
  senderUsername: zod.string(),
  recipientUsername: zod.string(),
  subject: zod.string(),
  body: zod.string(),
  isRead: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a message
 */
export const DeleteMessageParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Submit a staff application
 */
export const SubmitStaffApplicationBody = zod.object({
  gamertag: zod.string(),
  games: zod.array(zod.string()),
  experience: zod.string(),
  motivation: zod.string(),
});

/**
 * @summary List staff applications (moderator/admin)
 */
export const ListStaffApplicationsResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  gamertag: zod.string(),
  games: zod.array(zod.string()),
  experience: zod.string(),
  motivation: zod.string(),
  status: zod.string(),
  reviewNote: zod.string().nullable(),
  reviewedBy: zod.number().nullable(),
  createdAt: zod.string(),
});
export const ListStaffApplicationsResponse = zod.array(
  ListStaffApplicationsResponseItem,
);

/**
 * @summary Approve or reject a staff application (moderator/admin)
 */
export const ReviewStaffApplicationParams = zod.object({
  id: zod.coerce.number(),
});

export const ReviewStaffApplicationBody = zod.object({
  status: zod.string(),
  reviewNote: zod.string().optional(),
});

export const ReviewStaffApplicationResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  gamertag: zod.string(),
  games: zod.array(zod.string()),
  experience: zod.string(),
  motivation: zod.string(),
  status: zod.string(),
  reviewNote: zod.string().nullable(),
  reviewedBy: zod.number().nullable(),
  createdAt: zod.string(),
});