# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TAG - Tactical Adaptation Group API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthUser {
  id: number;
  username: string;
  email: string;
  role: string;
  status: string;
  /** @nullable */
  bio: string | null;
  /** @nullable */
  discordTag: string | null;
  createdAt: string;
}